Refactoring responsabilities
