+++ BGPViewerOpenApi/Service/AsProvider.cs
@@ -1,41 +1,29 @@
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using BGPViewerCore.Service;
 using BGPViewerCore.Model;
 using BGPViewerOpenApi.Model;
 
 namespace BGPViewerOpenApi.Service
 {
-    public class Provider
+    public class AsProvider
     {
-        private readonly IEnumerable<ApiBase> availableApis;
-        private readonly IServiceProvider serviceProvider;
+        private readonly ApiProvider apiProvider;
 
-        public Provider(IEnumerable<ApiBase> availableApis, IServiceProvider serviceProvider)
+        public AsProvider(ApiProvider apiProvider)
         {
-            this.availableApis = availableApis;
-            this.serviceProvider = serviceProvider;
-        }
-
-        internal async Task<IEnumerable<ApiBase>> ListAvailableAsync()
-        {
-            if(ContainsDuplicatedId()) throw new InvalidOperationException("There are one or more APIs with same ID.");
-
-            return await Task.FromResult(availableApis);
+            this.apiProvider = apiProvider;
         }
 
         internal async Task<AsnDetailsModel> GetDetailsAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             return await Task.FromResult(api.GetAsnDetails(asNumber));
         }
 
         internal async Task<Peers> GetPeersAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             var peersTuple = await Task.FromResult(api.GetAsnPeers(asNumber));
 
@@ -47,7 +35,7 @@ namespace BGPViewerOpenApi.Service
 
         internal async Task<Peers> GetUpstreamsAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             var upstreamsTuple = await Task.FromResult(api.GetAsnUpstreams(asNumber));
 
@@ -59,7 +47,7 @@ namespace BGPViewerOpenApi.Service
 
         internal async Task<Peers> GetDownstreamsAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             var downstreamsTuple = await Task.FromResult(api.GetAsnDownstreams(asNumber));
 
@@ -71,31 +59,16 @@ namespace BGPViewerOpenApi.Service
 
         internal async Task<IEnumerable<IxModel>> GetIxsAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             return await Task.FromResult(api.GetAsnIxs(asNumber));
         }
 
         internal async Task<AsnPrefixesModel> GetPrefixesAsync(int apiId, int asNumber)
         {
-            var api = GetApiById(apiId);
+            var api = apiProvider.GetApiById(apiId);
 
             return await Task.FromResult(api.GetAsnPrefixes(asNumber));
-        }
-
-        private IBGPViewerService GetApiById(int apiId)
-        {
-            var selectedApi = availableApis.FirstOrDefault(x => x.Id == apiId);
-            
-            if(selectedApi == null) throw new KeyNotFoundException($"Don't exist an API with ID {apiId}.");
-
-            return serviceProvider.GetService(selectedApi.ApiType) as IBGPViewerService;
-        }
-
-        private bool ContainsDuplicatedId()
-        {
-            var ids = availableApis.Select(x => x.Id);
-            return ids.Count() != ids.Distinct().Count();
         }
     }
 }
